refactor: put GeoJson under the right namespace
